backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
